% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notion_token_exists}
\alias{notion_token_exists}
\title{Check if Notion token is set}
\usage{
notion_token_exists()
}
\value{
\code{TRUE} if the token exists, \code{FALSE} otherwise.
}
\description{
Checks if the \code{NOTION_TOKEN} environment variable is set.
}
\examples{
notion_token_exists()
}
