% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notionapi-package.R
\docType{package}
\name{notionapi-package}
\alias{notionapi}
\alias{notionapi-package}
\title{notionapi: Client for the Notion API}
\description{
Enable programmatic interaction with 'Notion' pages, databases, blocks, comments, and users through the 'Notion API'. Provides both synchronous and asynchronous client interfaces for building workflows and automations that integrate with 'Notion' workspaces. Supports all 'Notion API' endpoints including content creation, data retrieval, and workspace management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brenwin1.github.io/notionapi/}
  \item \url{https://github.com/brenwin1/notionapi}
}

}
\author{
\strong{Maintainer}: Brenwin Ang \email{brenwinalj@gmail.com} [copyright holder]

}
\keyword{internal}
