% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.notion_response}
\alias{print.notion_response}
\title{Print the result of a Notion API call}
\usage{
\method{print}{notion_response}(x, ...)
}
\arguments{
\item{x}{The result object.}

\item{...}{Ignored.}
}
\value{
The JSON result.
}
\description{
Print the result of a Notion API call
}
