% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcr_example_helpers}
\alias{vcr_example_helpers}
\alias{vcr_example_start}
\alias{vcr_example_end}
\title{VCR Example Helpers}
\usage{
vcr_example_start(name)

vcr_example_end()
}
\arguments{
\item{name}{Character. The name of the cassette to be used.}
}
\description{
Internal functions for managing VCR cassettes in package examples.
These functions are not intended for direct use by package users.
}
\keyword{internal}
