% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{ViewsQueriesEndpoint}
\alias{ViewsQueriesEndpoint}
\title{R6 Class for Views Queries Endpoint}
\value{
A list containing the parsed API response.
}
\description{
Handle all views queries operations in the Notion API

\strong{Note:} Access this endpoint through the client instance, e.g., \code{notion$views$queries}. Not to be instantiated directly.
}
\examples{
\dontshow{if (notion_token_exists()) withAutoprint(\{ # examplesIf}
notion <- notion_client()
\dontshow{notionapi::vcr_example_start("notion-views-queries-create")}
# ----- Create a view query
notion$views$queries$create("33f33ea0-c1e4-8145-9555-000c19cc8d71")
\dontshow{notionapi::vcr_example_end()}
\dontshow{notionapi::vcr_example_start("notion-views-queries-results")}
# ----- Get view query results
notion$views$queries$results(
  "33f33ea0-c1e4-8145-9555-000c19cc8d71",
  "ee90464c-bd8d-4882-bba6-792a1dba18c6"
)
\dontshow{notionapi::vcr_example_end()}
\dontshow{notionapi::vcr_example_start("notion-views-queries-delete")}
# ----- Delete a view query
notion$views$queries$delete(
  "33f33ea0-c1e4-8145-9555-000c19cc8d71",
  "ee90464c-bd8d-4882-bba6-792a1dba18c6"
)
\dontshow{notionapi::vcr_example_end()}
\dontshow{\}) # examplesIf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ViewsQueriesEndpoint-new}{\code{ViewsQueriesEndpoint$new()}}
\item \href{#method-ViewsQueriesEndpoint-create}{\code{ViewsQueriesEndpoint$create()}}
\item \href{#method-ViewsQueriesEndpoint-results}{\code{ViewsQueriesEndpoint$results()}}
\item \href{#method-ViewsQueriesEndpoint-delete}{\code{ViewsQueriesEndpoint$delete()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ViewsQueriesEndpoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-ViewsQueriesEndpoint-new}{}}}
\subsection{Method \code{new()}}{
Initialise pages properties endpoint.
Not to be called directly, e.g., use \code{notion$views$queries} instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ViewsQueriesEndpoint$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Notion Client instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ViewsQueriesEndpoint-create"></a>}}
\if{latex}{\out{\hypertarget{method-ViewsQueriesEndpoint-create}{}}}
\subsection{Method \code{create()}}{
Create a view query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ViewsQueriesEndpoint$create(view_id, page_size = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{view_id}}{Character (required). The ID of the view.}

\item{\code{page_size}}{Integer. Number of items to return per page (1-100). Defaults to 100.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\href{https://developers.notion.com/reference/create-view-query}{Endpoint documentation}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ViewsQueriesEndpoint-results"></a>}}
\if{latex}{\out{\hypertarget{method-ViewsQueriesEndpoint-results}{}}}
\subsection{Method \code{results()}}{
Get view query results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ViewsQueriesEndpoint$results(
  view_id,
  query_id,
  start_cursor = NULL,
  page_size = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{view_id}}{Character (required). The ID of the view.}

\item{\code{query_id}}{Character (required). The ID of the query.}

\item{\code{start_cursor}}{Character. For pagination. If provided, returns results starting from this cursor.
If NULL, returns the first page of results.}

\item{\code{page_size}}{Integer. Number of items to return per page (1-100). Defaults to 100}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\href{https://developers.notion.com/reference/get-view-query-results}{Endpoint documentation}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ViewsQueriesEndpoint-delete"></a>}}
\if{latex}{\out{\hypertarget{method-ViewsQueriesEndpoint-delete}{}}}
\subsection{Method \code{delete()}}{
Delete a view query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ViewsQueriesEndpoint$delete(view_id, query_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{view_id}}{Character (required). The ID of the view.}

\item{\code{query_id}}{Character (required). The ID of the query.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\href{https://developers.notion.com/reference/delete-view-query}{Endpoint documentation}
}

}
}
