% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{DatabasesEndpoint}
\alias{DatabasesEndpoint}
\title{R6 Class for Databases Endpoint}
\value{
A list containing the parsed API response.
}
\description{
Handle all databases operations in the Notion API

\strong{Note:} Access this endpoint through the client instance, e.g., \code{notion$databases}. Not to be instantiated directly.
}
\examples{
\dontshow{if (notion_token_exists()) withAutoprint(\{ # examplesIf}
notion <- notion_client()
\dontshow{notionapi::vcr_example_start("notion-databases-create")}
# ----- Create a database
notion$databases$create(
  list(
    type = "page_id",
    page_id = "22f33ea0c1e480b99c77d1ab72aedff9"
  ),
  title = list(list(
    text = list(
      content = "Test Database"
    )
  ))
)
\dontshow{notionapi::vcr_example_end()}
\dontshow{notionapi::vcr_example_start("notion-databases-retrieve")}
# ----- Retrieve a database
notion$databases$retrieve("efab1bec-0094-4afe-a90c-c9b72d538b4b")
\dontshow{notionapi::vcr_example_end()}
\dontshow{notionapi::vcr_example_start("notion-database-update")}
# ----- Update a database
notion$databases$update(
  "efab1bec-0094-4afe-a90c-c9b72d538b4b",
  description = list(list(
    text = list(
      content = "For testing purposes"
    )
  )),
  icon = list(
    emoji = "🤩"
  )
)
\dontshow{notionapi::vcr_example_end()}
\dontshow{\}) # examplesIf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabasesEndpoint-new}{\code{DatabasesEndpoint$new()}}
\item \href{#method-DatabasesEndpoint-create}{\code{DatabasesEndpoint$create()}}
\item \href{#method-DatabasesEndpoint-retrieve}{\code{DatabasesEndpoint$retrieve()}}
\item \href{#method-DatabasesEndpoint-update}{\code{DatabasesEndpoint$update()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabasesEndpoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabasesEndpoint-new}{}}}
\subsection{Method \code{new()}}{
Initialise databases endpoint.
Not to be called directly, e.g., use \code{notion$databases} instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabasesEndpoint$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Notion Client instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabasesEndpoint-create"></a>}}
\if{latex}{\out{\hypertarget{method-DatabasesEndpoint-create}{}}}
\subsection{Method \code{create()}}{
Create a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabasesEndpoint$create(
  parent,
  title = NULL,
  description = NULL,
  is_inline = NULL,
  initial_data_source = NULL,
  icon = NULL,
  cover = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{Named list (JSON object) (required). The parent page or workspace
where the database will be created.}

\item{\code{title}}{List of lists (JSON array). The title of the database.}

\item{\code{description}}{List of lists (JSON array). The description of the database.}

\item{\code{is_inline}}{Boolean. Whether the database should be displayed inline
in the parent page. Defaults to false.}

\item{\code{initial_data_source}}{Named list (JSON object). Initial data source
configuration for the database}

\item{\code{icon}}{Named list (JSON object). The icon for the database.}

\item{\code{cover}}{Named list (JSON object). The cover image for the database.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\href{https://developers.notion.com/reference/create-database}{Endpoint documentation}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabasesEndpoint-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-DatabasesEndpoint-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabasesEndpoint$retrieve(database_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{String (required). The ID of a Notion database.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\href{https://developers.notion.com/reference/retrieve-database}{Endpoint documentation}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabasesEndpoint-update"></a>}}
\if{latex}{\out{\hypertarget{method-DatabasesEndpoint-update}{}}}
\subsection{Method \code{update()}}{
Update a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabasesEndpoint$update(
  database_id,
  parent = NULL,
  title = NULL,
  description = NULL,
  is_inline = NULL,
  icon = NULL,
  cover = NULL,
  in_trash = NULL,
  is_locked = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{String (required). The ID of a Notion database.}

\item{\code{parent}}{Named list (JSON object). The parent page or workspace
to move the database to. If not provided, the database will not be moved.}

\item{\code{title}}{List of lists (JSON array). The updated title of the database.}

\item{\code{description}}{List of lists (JSON array). The updated description of the database.}

\item{\code{is_inline}}{Boolean. Whether the database should be displayed in the
parent page.}

\item{\code{icon}}{Named list (JSON object). The updated icon for the database.}

\item{\code{cover}}{Named list (JSON object). The updated cover image for the database.}

\item{\code{in_trash}}{Boolean. Whether the database should be moved to or from the trash.}

\item{\code{is_locked}}{Boolean. Whether the database should be locked from editing.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\href{https://developers.notion.com/reference/update-database}{Endpoint documentation}
}

}
}
