% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{no_config}
\alias{no_config}
\title{Create empty property configuration}
\usage{
no_config()
}
\value{
An empty \verb{named list} that serialises to \code{{}} in JSON
}
\description{
Helper function that creates an empty named list for property configurations
that require no additional settings.

Many \href{https://developers.notion.com/reference/property-object}{database properties} like text, checkbox and date
do not need configuration settings.
This function returns the empty configuration (\code{{}} in JSON)
that these properties expect.
}
