% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notionapi-package.R
\docType{package}
\name{notionapi-package}
\alias{notionapi}
\alias{notionapi-package}
\title{notionapi: R Client for the Notion API}
\description{
Provides an R interface to the Notion API, enabling users to programatically interact with Notion pages, databases, blocks, comments, and users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brenwin1.github.io/notionapi/}
  \item \url{https://github.com/brenwin1/notionapi}
}

}
\author{
\strong{Maintainer}: Brenwin Ang \email{brenwinalj@gmail.com} [copyright holder]

}
\keyword{internal}
